Move model-editor to model
