+++ src/app/root/[domain]/[lang]/model/dimensions/[dimensionId]/page.tsx
