+++ src/app/root/[domain]/[lang]/model/layout.tsx
