+++ src/app/root/[domain]/[lang]/model/nodes/page.tsx
