+++ src/components/model-editor/ModelEditorNav.tsx
@@ -119,42 +119,42 @@ const TABS: TabDef[] = [
   {
     label: 'Model',
     matches: (path) =>
-      !path.includes('/model-editor/nodes') &&
-      !path.includes('/model-editor/datasets') &&
-      !path.includes('/model-editor/dimensions'),
+      !path.includes('/model/nodes') &&
+      !path.includes('/model/datasets') &&
+      !path.includes('/model/dimensions'),
     href: '',
     Icon: House,
   },
   {
     label: 'Nodes',
-    matches: (path) => path.includes('/model-editor/nodes'),
+    matches: (path) => path.includes('/model/nodes'),
     href: '/nodes',
     Icon: Diagram2,
   },
   {
     label: 'Datasets',
-    matches: (path) => path.includes('/model-editor/datasets'),
+    matches: (path) => path.includes('/model/datasets'),
     href: '/datasets',
     Icon: Database,
   },
   {
     label: 'Dimensions',
-    matches: (path) => path.includes('/model-editor/dimensions'),
+    matches: (path) => path.includes('/model/dimensions'),
     href: '/dimensions',
     Icon: BoxIcon,
   },
 ];
 
 function getModelEditorBase(pathname: string): string {
-  const idx = pathname.indexOf('/model-editor');
-  return idx >= 0 ? pathname.slice(0, idx) + '/model-editor' : '/model-editor';
+  const idx = pathname.indexOf('/model');
+  return idx >= 0 ? pathname.slice(0, idx) + '/model' : '/model';
 }
 
 type SearchMode = 'nodes' | 'datasets' | 'dimensions';
 
 function getSearchMode(pathname: string): SearchMode {
-  if (pathname.includes('/model-editor/datasets')) return 'datasets';
-  if (pathname.includes('/model-editor/dimensions')) return 'dimensions';
+  if (pathname.includes('/model/datasets')) return 'datasets';
+  if (pathname.includes('/model/dimensions')) return 'dimensions';
   return 'nodes';
 }
