+++ src/components/model-editor/dimensions/DimensionEditor.tsx
@@ -76,10 +76,8 @@ function toRows(dim: InstanceDimensionFieldsFragment): CategoryRow[] {
 }
 
 function getListBase(pathname: string): string {
-  const idx = pathname.indexOf('/model-editor');
-  return idx >= 0
-    ? pathname.slice(0, idx) + '/model-editor/dimensions'
-    : '/model-editor/dimensions';
+  const idx = pathname.indexOf('/model');
+  return idx >= 0 ? pathname.slice(0, idx) + '/model/dimensions' : '/model/dimensions';
 }
 
 function genUuid(): string {
